requirements.txt
python-dotenv
openai
tenacity
tiktoken
adversarial-robustness-toolbox